Refactor to fix SonarCloud warnings and improve code quality.
