+++ Src/CrispyWaffle.Redis/Log/PropagationStrategy/ProcessPropagationStrategy.cs
@@ -6,17 +6,28 @@ using StackExchange.Redis;
 namespace CrispyWaffle.Redis.Log.PropagationStrategy
 {
     /// <summary>
-    /// The process propagation strategy
+    /// Represents a process-based propagation strategy for publishing messages to Redis.
     /// </summary>
-    /// <seealso cref="IPropagationStrategy" />
+    /// <remarks>
+    /// This class implements the <see cref="IPropagationStrategy"/> interface and defines how messages
+    /// are propagated to a Redis publisher. It uses the process ID as part of the queue naming convention
+    /// to uniquely identify messages for the current process.
+    /// </remarks>
+    /// <seealso cref="IPropagationStrategy"/>
     public sealed class ProcessPropagation : IPropagationStrategy
     {
         /// <summary>
-        /// Propagates the specified message using specific strategy to the publisher.
+        /// Propagates the specified message to the Redis publisher using the process-based strategy.
         /// </summary>
-        /// <param name="message">The message.</param>
-        /// <param name="queuePrefix">The queue prefix.</param>
-        /// <param name="publisher">The publisher.</param>
+        /// <param name="message">The message to propagate. This will be published to the Redis channel.</param>
+        /// <param name="queuePrefix">The prefix to be used in the Redis queue names. It is combined with the process ID to form the final queue name.</param>
+        /// <param name="publisher">The <see cref="ISubscriber"/> responsible for publishing the message to Redis.</param>
+        /// <remarks>
+        /// This method will first publish the process ID to the Redis channel formatted as
+        /// "<paramref name="queuePrefix"/>-queues". Then, it will publish the message to the
+        /// channel "<paramref name="queuePrefix"/>-<see cref="EnvironmentHelper.ProcessId"/>", where
+        /// <see cref="EnvironmentHelper.ProcessId"/> is the unique process ID for the current process.
+        /// </remarks>
         public void Propagate(string message, string queuePrefix, ISubscriber publisher)
         {
             publisher.Publish($"{queuePrefix}-queues", EnvironmentHelper.ProcessId);
@@ -24,13 +35,19 @@ namespace CrispyWaffle.Redis.Log.PropagationStrategy
         }
 
         /// <summary>
-        /// propagate as an asynchronous operation.
+        /// Asynchronously propagates the specified message to the Redis publisher using the process-based strategy.
         /// </summary>
-        /// <param name="message">The message.</param>
-        /// <param name="queuePrefix">The queue prefix.</param>
-        /// <param name="publisher">The publisher.</param>
-        /// <param name="cancellationToken">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>Task.</returns>
+        /// <param name="message">The message to propagate. This will be published to the Redis channel asynchronously.</param>
+        /// <param name="queuePrefix">The prefix to be used in the Redis queue names. It is combined with the process ID to form the final queue name.</param>
+        /// <param name="publisher">The <see cref="ISubscriber"/> responsible for publishing the message to Redis asynchronously.</param>
+        /// <param name="cancellationToken">A <see cref="CancellationToken"/> that can be used to cancel the operation.</param>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        /// <remarks>
+        /// This method will first asynchronously publish the process ID to the Redis channel formatted as
+        /// "<paramref name="queuePrefix"/>-queues". Then, it will asynchronously publish the message to the
+        /// channel "<paramref name="queuePrefix"/>-<see cref="EnvironmentHelper.ProcessId"/>, where
+        /// <see cref="EnvironmentHelper.ProcessId"/> is the unique process ID for the current process.
+        /// </remarks>
         public async Task PropagateAsync(
             string message,
             string queuePrefix,
@@ -46,6 +63,7 @@ namespace CrispyWaffle.Redis.Log.PropagationStrategy
             await publisher
                 .PublishAsync($"{queuePrefix}-queues", EnvironmentHelper.ProcessId)
                 .ConfigureAwait(false);
+
             await publisher
                 .PublishAsync($"{queuePrefix}-{EnvironmentHelper.ProcessId}", message)
                 .ConfigureAwait(false);
