+++ Tests/CrispyWaffle.Tests/Serialization/JsonSerializerAdapterTests.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using CrispyWaffle.Serialization;
 using CrispyWaffle.Serialization.Adapters;
 using CrispyWaffle.Serialization.SystemTextJson;
+using CrispyWaffle.Tests.Serialization.TestObjects_;
 using FluentAssertions;
 using Xunit;
