+++ Tests/CrispyWaffle.Tests/Serialization/NewtonsoftJsonSerializerAdapterTests.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using CrispyWaffle.Serialization.Adapters;
+using CrispyWaffle.Tests.Serialization.TestObjects_;
 using FluentAssertions;
 using Xunit;
