+++ Tests/CrispyWaffle.Tests/Serialization/SerializerFactoryTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using CrispyWaffle.Serialization;
+using CrispyWaffle.Tests.Serialization.TestObjects_;
 using Xunit;
 
 namespace CrispyWaffle.Tests.Serialization;
