+++ Tests/CrispyWaffle.Tests/Serialization/TestObjects_/SampleNonSerializableClass.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace CrispyWaffle.Tests.Serialization;
+namespace CrispyWaffle.Tests.Serialization.TestObjects_;
 
 [ExcludeFromCodeCoverage]
 public class SampleNonSerializableClass
