+++ Tests/CrispyWaffle.Tests/Serialization/TestObjects_/SampleXmlClass.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using CrispyWaffle.Serialization;
 
-namespace CrispyWaffle.Tests.Serialization;
+namespace CrispyWaffle.Tests.Serialization.TestObjects_;
 
 [Serializer]
 [ExcludeFromCodeCoverage]
@@ -11,14 +11,10 @@ public class SampleXmlClass : IEquatable<SampleXmlClass>
     public bool Equals(SampleXmlClass other)
     {
         if (ReferenceEquals(null, other))
-        {
             return false;
-        }
 
         if (ReferenceEquals(this, other))
-        {
             return true;
-        }
 
         return Code == other.Code
             && string.Equals(String, other.String, StringComparison.InvariantCultureIgnoreCase)
@@ -29,19 +25,13 @@ public class SampleXmlClass : IEquatable<SampleXmlClass>
     public override bool Equals(object obj)
     {
         if (ReferenceEquals(null, obj))
-        {
             return false;
-        }
 
         if (ReferenceEquals(this, obj))
-        {
             return true;
-        }
 
         if (obj.GetType() != GetType())
-        {
             return false;
-        }
 
         return Equals((SampleXmlClass)obj);
     }
