+++ appveyor.yml
@@ -1,4 +1,4 @@
-version: 8.2.{build}
+version: 9.0.{build}
 skip_tags: true
 image: Visual Studio 2022
 configuration: Release
